Translation data contain
On the day of the senior high school entrance examination, we didn't even say goodbye and left with our schoolbags on our backs. Some people had never seen each other since that day, just like the wind. We would never see each other again after graduation. The moment we walked out of the school gate, the sun was still shining, as if everything had not changed, but our hearts were empty. The once familiar classroom corridor turned into a three-year junior high school in a twinkling of an eye. I recorded every bit of us, and I was crazy with my friends, swearing every day, even if I was late, I went back to the classroom slowly. At that time, I complained about school every day, but when I passed the familiar road, I couldn't help looking at us again. At that time, we always felt that the future was still very long, and we always thought that goodbye was a matter of course. Later, I realized the place we were trying to escape. Is the most memorable place, in fact, we miss is not junior high school, but junior high school that group of people playing together when the joy, in fact, the teacher really did not lie to us, junior high school three years is really very short, but the reality tells me.Some of us can never meet again once we are separated. Those once smiling faces and those once agreements are gradually blurred over time and eventually disappear in the depths of memory.